Implemented user,comment and most of the post controllers

+++ Backend/utils/cloudinary.js
@@ -11,24 +11,30 @@ const uploadOnCloudinary=async(localFilePath)=>{
     try {
         if(!localFilePath){return null;}
         const upload=await cloudinary.uploader.upload(localFilePath,{
-            resource_type:auto
+            resource_type:"auto"
         })
+        if(fs.existsSync){
         fs.unlinkSync(localFilePath)
+        }
+
         return upload;
     } catch (error) {
+        if(fs.existsSync){
         fs.unlinkSync(localFilePath)
+        }
+        
         console.log("Cloudinary error",error)
     }
 }
 
-const deleteOnCloudinary=async(localFilePath)=>{
+const deleteOnCloudinary=async(publicId)=>{
     try {
-        if(!localFilePath){return;}
-        const delete_file=cloudinary.uploader.destroy(localFilePath);
+        if(!publicId){return;}
+        const delete_file=await cloudinary.uploader.destroy(publicId);
         return delete_file
     } catch (error) {
         console.log("Delete failed")
     }
 }
 
-export {uploadOnCloudinary}
+export {uploadOnCloudinary,deleteOnCloudinary}
